Restore static dashboard page and relocate CSS assets
