+++ src/css/AssistantPanel.css
